requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.3
openpyxl==3.1.5
pandas==3.0.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
